+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - update core to state template 5.0.7
 - update dependency versions
+- add unit tests (Layout)
 
 ## v0.3.2 (2019-06-04)